% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaZING_Model.R
\name{BaZING_Model}
\alias{BaZING_Model}
\title{BaZING_Model Function
This function implements the BaZING model for microbiome data analysis.}
\usage{
BaZING_Model(
  formatted_data,
  covar,
  x,
  n.chains = 3,
  n.adapt = 100,
  n.iter.update = 5,
  n.iter.coda = 5
)
}
\arguments{
\item{formatted_data}{An object containing formatted microbiome data.}

\item{covar}{A vector of covariates.}

\item{x}{A vector of exposures.}

\item{n.adapt}{the number of iterations for adaptation in jags.model function}

\item{n.iter.update}{number of iterations in update function}

\item{n.iter.coda}{number of iterations in coda.samples function}

\item{n.chain}{the number of parallel chains for the model in jags.model function}
}
\value{
A data frame containing results of the Bayesian analysis.
}
\description{
Load Packages
}
