% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_BaHZING.R
\name{BHMRmicrobiome}
\alias{BHMRmicrobiome}
\alias{Format_BaHZING}
\title{Format_BaHZING Function}
\usage{
Format_BaHZING(phyloseq.object)
}
\arguments{
\item{phyloseq.object}{A phyloseq object.}
}
\value{
A list with the following elements:
\itemize{
\item \code{Table}: Formatted microbiome data as a data frame.
\item \code{Species.Genus.Matrix}: Data frame for species-genus relationships (optional).
\item \code{Genus.Family.Matrix}: Data frame for genus-family relationships (optional).
\item \code{Family.Order.Matrix}: Data frame for family-order relationships (optional).
\item \code{Order.Class.Matrix}: Data frame for order-class relationships (optional).
\item \code{Class.Phylum.Matrix}: Data frame for class-phylum relationships (optional).
}
}
\description{
This function takes a phyloseq object and performs formatting operations on it,
including modifying the taxonomic table, uniting taxonomic levels, and creating
matrices based on taxonomic information.
}
\details{
The column names 'Kingdom', 'Phylum', 'Class', 'Order', 'Family', 'Genus', and 'Species'
in the tax_table of the phyloseq object should be user-defined and assigned in this function.
The function will use these column names to perform various operations.
}
\note{
The column names 'Kingdom', 'Phylum', 'Class', 'Order', 'Family', 'Genus', and 'Species' in the
tax_table are expected to be user-defined and assigned within the function.
}
\keyword{R}
\keyword{data}
\keyword{formatting}
\keyword{microbiome,}
\keyword{package,}
\keyword{phyloseq,}
\keyword{table,}
\keyword{taxonomic}
