% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHMP.R
\docType{data}
\name{iHMP}
\alias{iHMP}
\title{iHMP data}
\format{
\subsection{\code{iHMP}}{

A phyloseq object with microbiome data for 105 participants at their
first visit.
}
}
\source{
https://www.nature.com/articles/s41586-019-1237-9>
}
\usage{
iHMP
}
\description{
A subset of data from the Integrative Human Microbiome Project
}
\keyword{datasets}
