% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHMP_Reduced_update.R
\docType{data}
\name{iHMP_Reduced_update}
\alias{iHMP_Reduced_update}
\title{iHMP_Reduced_update data}
\format{
\subsection{\code{iHMP_Reduced_update}}{

A phyloseq object with subset microbiome data for 105 participants at their
first visit.
}
}
\source{
https://www.nature.com/articles/s41586-019-1237-9>
}
\usage{
iHMP_Reduced_update
}
\description{
A subset of data from the Integrative Human Microbiome Project
}
\keyword{datasets}
