% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP_qgcomp.R
\name{ZIP_qgcomp}
\alias{ZIP_qgcomp}
\title{ZIP_qgcomp Function
This function implements the BaHZING model for microbiome data analysis.}
\usage{
ZIP_qgcomp(formatted_data, x, covar = NULL, q = NULL, verbose = TRUE)
}
\arguments{
\item{formatted_data}{An object containing formatted microbiome data.}

\item{x}{A vector of column names of the exposures.}

\item{covar}{An optional vector of the column names of covariates.}

\item{q}{An integer specifying the number of quantiles.}

\item{verbose}{If TRUE (default), function returns information a data quality
check.}
}
\value{
A data frame containing results of the qgcomp analysis, with the
following columns:
\itemize{
\item taxa_full: Full Taxa information, including all levels of the taxonomy.
Taxanomic levels are split by two underscores ('__').
\item taxa_name: Taxa name, which is the last level of the taxonomy.
\item domain: domain of the taxa.
\item exposure: Exposure name (either one of  the individual exposures, or the
mixture).
\item estimate: Point estimate or mixture estimate.
\item lcl: 95\% Interval Lower Limit. Calculated as estimate - 1.96 × standard error
\item ucl: 95\% Interval Upper Limit. Calculated as estimate + 1.96 × standard error.
\item p_value: P-value for the hypothesis that the effect estimate equals zero.
\item model: Indicates the method used in the qgcomp() analysis.
\itemize{
\item "Poisson": Poisson regression model (from qgcomp()).
}
}
}
\description{
ZIP_qgcomp Function
This function implements the BaHZING model for microbiome data analysis.
}
