% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaHZING_Model.R
\name{BaHZING_Model}
\alias{BaHZING_Model}
\title{BaHZING_Model Function
This function implements the BaHZING model for microbiome data analysis.}
\arguments{
\item{formatted_data}{An object containing formatted microbiome data.}

\item{covar}{A vector of covariates.}

\item{x}{A vector of exposures.}

\item{n.chains}{An integer specifying the number of parallel chains for the model in jags.model function}

\item{n.adapt}{An integer specifying the number of iterations for adaptation in jags.model function}

\item{n.burnin}{An integer specifying number of iterations in update function}

\item{n.sample}{An integer specifying the number of iterations in coda.samples function}

\item{exposure_standardization}{"standard_normal" or "quantile"}

\item{counterfactual_profiles}{A 2xP matrix or a vector with length of 2; P is the column number of exposure dataframe.}

\item{q}{An integer specifying the number of quantile groups}
}
\value{
A data frame containing results of the Bayesian analysis.
}
\description{
Load Packages
}
